backend/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
slowapi==0.1.9
tabulate==0.9.0

# Dev tools
black==24.8.0
flake8==7.1.1
